src/requirements.txt
numpy>=1.22.4,<1.22.5
scipy>=1.10.1,<1.10.2
matplotlib>=3.7.1,<3.7.2
attrs>=23.1.0,<23.1.1
tqdm>=4.65.0,<4.65.1
pyqt5>=5.15.9,<5.15.10
sympy>=1.12, <1.13
